Change AUTH token to have length of 20.
- Fix authe token request to use U64 for activation token

+++ app/data/authentication_token_request_message.rb
@@ -4,6 +4,6 @@ class AuthenticationTokenRequestMessage < BinData::Record
 
   endian ENDIAN
 
-  uint48 :activation_token
+  uint64 :activation_token
 end
